+++ Finjector.Web/ClientApp/src/pages/Teams/Team.tsx
@@ -3,26 +3,9 @@ import { SearchBar } from "../../components/Shared/SearchBar";
 import { useGetTeam } from "../../queries/teamQueries";
 import { useParams } from "react-router-dom";
 import FolderList from "../../components/Teams/FolderList";
-import FinLoader from "../../components/Shared/LoadingAndErrors/FinLoader";
-import DeleteTeamModal from "../../components/Teams/DeleteTeamModal";
-import LeaveTeamModal from "../../components/Teams/LeaveTeamModal";
-import { isPersonalOrDefault } from "../../util/teamDefinitions";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faPlus,
-  faUsers,
-  faPencil,
-  faUserTie,
-  faPersonThroughWindow,
-  faTrash,
-} from "@fortawesome/free-solid-svg-icons";
-import FinButton from "../../components/Shared/FinButton";
-import PageTitle from "../../components/Shared/Layout/PageTitle";
 import PageBody from "../../components/Shared/Layout/PageBody";
-import PageInfo from "../../components/Shared/Layout/PageInfo";
-import FinButtonDropdown from "../../components/Shared/FinButtonDropdown";
-import FinButtonDropdownItem from "../../components/Shared/FinButtonDropdownItem";
+import { FinQueryStatus } from "../../types";
+import TeamTitle from "../../components/Teams/TeamTitle";
 
 const Team: React.FC = () => {
   // get id from url
@@ -30,128 +13,32 @@ const Team: React.FC = () => {
 
   const [search, setSearch] = React.useState("");
 
-  const [modalOpen, setModalOpen] = React.useState("");
-  const toggleModal = (modalType: string) => {
-    setModalOpen(modalType);
+  const teamModelQuery = useGetTeam(teamId);
+
+  const queryStatus: FinQueryStatus = {
+    isError: teamModelQuery.isError,
+    isInitialLoading: teamModelQuery.isInitialLoading,
+    error: teamModelQuery.error,
   };
-
-  const teamModel = useGetTeam(teamId);
-
-  if (teamModel.isLoading) {
-    return <FinLoader />;
-  }
-
-  if (teamModel.error) {
-    return (
-      <div>
-        Something went wrong. Ensure you have permission to view this team.
-      </div>
-    );
-  }
-
-  const isTeamAdmin = teamModel.data?.team.myTeamPermissions.some(
-    (p) => p === "Admin"
-  );
-
-  const limitedTeam = isPersonalOrDefault(teamModel.data?.team.name);
 
   return (
     <div>
-      <PageTitle>
-        <div className="col-12 col-md-9">
-          <h1>{teamModel.data?.team.name}</h1>
-        </div>
-        <div className="col-12 col-md-3 text-end">
-          <FinButtonDropdown shouldRenderAsDropdown={!limitedTeam}>
-            {/* don't show team admins if you are an admin or if it's a personal team */}
-            {limitedTeam ||
-              (!isTeamAdmin && (
-                <FinButtonDropdownItem>
-                  <FinButton borderless={true} to={`/teams/${teamId}/admins`}>
-                    <FontAwesomeIcon icon={faUserTie} />
-                    View Team Admins
-                  </FinButton>
-                </FinButtonDropdownItem>
-              ))}
-            {!limitedTeam && isTeamAdmin && (
-              <>
-                <FinButtonDropdownItem>
-                  <FinButton
-                    borderless={true}
-                    to={`/teams/${teamId}/folders/create`}
-                  >
-                    <FontAwesomeIcon icon={faPlus} />
-                    Create New Folder
-                  </FinButton>
-                </FinButtonDropdownItem>
-                <FinButtonDropdownItem>
-                  <FinButton borderless={true} to={`/teams/${teamId}/edit`}>
-                    <FontAwesomeIcon icon={faPencil} />
-                    Edit Team
-                  </FinButton>
-                </FinButtonDropdownItem>
-                <FinButtonDropdownItem>
-                  <FinButton
-                    borderless={true}
-                    to={`/teams/${teamId}/permissions`}
-                  >
-                    <FontAwesomeIcon icon={faUsers} />
-                    Manage Team Users
-                  </FinButton>
-                </FinButtonDropdownItem>
-                <FinButtonDropdownItem>
-                  <FinButton
-                    borderless={true}
-                    onClick={() => toggleModal("delete")}
-                  >
-                    <FontAwesomeIcon icon={faTrash} />
-                    Delete Team
-                  </FinButton>{" "}
-                </FinButtonDropdownItem>
-              </>
-            )}
-
-            {!limitedTeam && (
-              <FinButtonDropdownItem>
-                <FinButton
-                  borderless={true}
-                  onClick={() => toggleModal("leave")}
-                >
-                  <FontAwesomeIcon icon={faPersonThroughWindow} />
-                  Leave Team
-                </FinButton>
-              </FinButtonDropdownItem>
-            )}
-          </FinButtonDropdown>
-        </div>
-      </PageTitle>
-      <PageInfo>{teamModel.data?.team.description}</PageInfo>
+      <TeamTitle
+        teamModelData={teamModelQuery.data}
+        queryStatus={queryStatus}
+        teamId={teamId}
+      />
       <PageBody>
-        {!limitedTeam && (
-          <LeaveTeamModal
-            teamId={teamId}
-            isAdmin={isTeamAdmin !== undefined && isTeamAdmin}
-            isOpen={modalOpen === "leave"}
-            closeModal={() => toggleModal("")}
-          />
-        )}
-        {!limitedTeam && isTeamAdmin && (
-          <DeleteTeamModal
-            teamId={teamId}
-            isOpen={modalOpen === "delete"}
-            closeModal={() => toggleModal("")}
-          />
-        )}
         <SearchBar
           placeholderText={
-            !!teamModel.data?.team.name
-              ? `Search Within ${teamModel.data?.team.name}`
+            !!teamModelQuery.data?.team.name
+              ? `Search Within ${teamModelQuery.data?.team.name}`
               : "Search My Teams"
           }
           search={search}
           setSearch={setSearch}
         />
-        <FolderList teamModel={teamModel.data} filter={search} />
+        <FolderList teamModel={teamModelQuery.data} filter={search} />
       </PageBody>
     </div>
   );
